project/drone/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-celery-results==2.5.0
djangorestframework==3.14.0
dnspython==2.3.0
eventlet==0.33.3
filelock==3.10.6
filterpy==1.4.5
fonttools==4.39.2
greenlet==2.0.2
idna==3.4
imageio==2.26.1
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prompt-toolkit==3.0.38
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
tifffile==2023.3.21
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
